requirements.txt
openai
numpy
pandas
azure-identity
azure-search-documents
python-dotenv
pillow
promptflow-evals
ipykernel